test: extract examples test data preparation to own helper
